Исправление ошибок Вынесена общее начало API в общую константу API приведено к общему виду
