+++ src/main/java/com/tbank/ttravels_backend/controller/TransferController.java
@@ -10,7 +10,7 @@ import org.springframework.web.bind.annotation.*;
 
 
 @RestController
-@RequestMapping("/api/transfers")
+@RequestMapping("/transfers")
 @RequiredArgsConstructor
 public class TransferController {
