+++ src/main/java/com/tbank/ttravels_backend/service/AccountService.java
@@ -1,4 +1,4 @@
-package com.tbank.ttravels_backend.security;
+package com.tbank.ttravels_backend.service;
 
 import com.tbank.ttravels_backend.dto.auth.AccountResponse;
 import com.tbank.ttravels_backend.dto.auth.AuthResponse;
@@ -14,6 +14,8 @@ import com.tbank.ttravels_backend.exception.UserNotFoundException;
 import com.tbank.ttravels_backend.repository.PasswordCredentialRepository;
 import com.tbank.ttravels_backend.repository.RefreshTokenRepository;
 import com.tbank.ttravels_backend.repository.UserRepository;
+import com.tbank.ttravels_backend.security.TokenHashService;
+import com.tbank.ttravels_backend.security.TokenIssuer;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.crypto.password.PasswordEncoder;
